gh-985 parse valid geo range filter

+++ adapters/handlers/rest/embedded_spec.go
@@ -3654,7 +3654,7 @@ func init() {
           "properties": {
             "max": {
               "type": "number",
-              "format": "float"
+              "format": "float64"
             }
           }
         },
@@ -7364,7 +7364,7 @@ func init() {
           "properties": {
             "max": {
               "type": "number",
-              "format": "float"
+              "format": "float64"
             }
           }
         },
